requirements.txt
mutagen
requests
eyed3
python-dotenv
